% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_confidence.R
\name{calibrate_confidence}
\alias{calibrate_confidence}
\title{Calibrate Confidence Scores}
\usage{
calibrate_confidence(scores)
}
\arguments{
\item{scores}{Numeric vector}
}
\value{
Numeric vector
}
\description{
Converts raw confidence scores
into calibrated percentile scores.
}
\examples{
scores <- c(
  0.2,
  0.5,
  0.8
)

calibrate_confidence(
  scores
)

}
