% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_confidence.R
\name{classify_confidence}
\alias{classify_confidence}
\title{Classify Confidence Levels}
\usage{
classify_confidence(scores)
}
\arguments{
\item{scores}{Numeric vector}
}
\value{
Character vector
}
\description{
Converts confidence scores into
confidence classes.
}
\examples{
scores <- c(
  0.2,
  0.7,
  0.9
)

classify_confidence(
  scores
)

}
