% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_labels.R
\name{match_labels}
\alias{match_labels}
\title{Ontology Label Matching}
\usage{
match_labels(label, marker_names)
}
\arguments{
\item{label}{Character label}

\item{marker_names}{Character vector}
}
\value{
Character vector
}
\description{
Matches predicted labels
to marker database labels.
}
\examples{
match_labels(
  "T_cells",
  c(
    "T_cell",
    "B_cell"
  )
)

}
